Moved typed inbox to testkit #22753
